CoinFlow-Telegram-Crypto-Bot/pyproject.toml
[tool.poetry]
name = "coinflow"
version = "3.1.0"
description = "Advanced Telegram bot for currency conversion with 50+ cryptocurrencies and AI-powered features"
authors = ["bobberdolle1"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "*"
python-telegram-bot = "*"
python-dotenv = "*"
matplotlib = "*"
yfinance = "*"
numpy = "*"
pandas = "*"
statsmodels = "*"
scikit-learn = "*"
apscheduler = "*"
bestchange-api = "*"
sqlalchemy = "*"
aiohttp = "*"
feedparser = "*"
plotly = "*"
prophet = {version = "*", optional = true}
google-auth = {version = "*", optional = true}
google-auth-oauthlib = {version = "*", optional = true}
google-api-python-client = {version = "*", optional = true}
notion-client = {version = "*", optional = true}
SpeechRecognition = {version = "*", optional = true}
pydub = {version = "*", optional = true}
faster-whisper = {version = "*", optional = true}
fastapi = {version = "*", optional = true}
uvicorn = {version = "*", optional = true}
jinja2 = {version = "*", optional = true}
python-multipart = {version = "*", optional = true}

[tool.poetry.extras]
sheets = ["google-auth", "google-auth-oauthlib", "google-api-python-client"]
notion = ["notion-client"]
voice = ["SpeechRecognition", "pydub"]
voice-fast = ["SpeechRecognition", "pydub", "faster-whisper"]
webapp = ["fastapi", "uvicorn", "jinja2", "python-multipart"]
prediction = ["prophet"]
all = ["google-auth", "google-auth-oauthlib", "google-api-python-client", "notion-client", "SpeechRecognition", "pydub", "fastapi", "uvicorn", "jinja2", "python-multipart", "prophet"]
all-fast = ["google-auth", "google-auth-oauthlib", "google-api-python-client", "notion-client", "SpeechRecognition", "pydub", "faster-whisper", "fastapi", "uvicorn", "jinja2", "python-multipart", "prophet"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"